titiler-app/requirements.txt
fastapi
uvicorn[standard]
titiler.core